Added ClinGenDosageSensitivity Parser
handle parsing of gene- and region-based info from clingen dosage sensitivity

add NCBIGeneParser (with pseudo-edges)

For now, this creates somewhat meaningless edges after normalization
because we provide "NCBIGene" is-same-as "HGNC", but both get
normalized to ncbi gene ids. This is a workaround to try to add
location attributes to gene entities.

Changed the url to fetch the data sources from ftp --> http

Update Parser